move deposit, proposal to bridge/
